+++ CHANGELOG.md
@@ -6,9 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 48.4.2 - 2020-10-02
 - [Patch] Updated **Tile** with new `unsavedChangesText` prop to indicate a changed item ([#1427](https://github.com/optimizely/oui/pull/1427))
 - [Patch] Revert! Memomized the **DraggableItem** props to increase re-render performance. ([#1426](https://github.com/optimizely/oui/pull/1426))
-
 
 ## 48.4.1 - 2020-10-01
 - [Patch] Updated  **Sidebar** with new prop `sticky` to allow it to maintain it's fixed position when needed. ([#1424](https://github.com/optimizely/oui/pull/1424))